# Use tiny_inflate instead of pako