--- a/bin/wasm-node/javascript/src/instance/raw-instance.ts
+++ b/bin/wasm-node/javascript/src/instance/raw-instance.ts
@@ -15,14 +15,13 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import pako from 'pako';
-
 import * as buffer from './buffer.js';
+import { default as tinyInflate } from '../tiny-inflate/index.js'
 
 import { Config as SmoldotBindingsConfig, default as smoldotLightBindingsBuilder } from './bindings-smoldot-light.js';
 import { Config as WasiConfig, default as wasiBindingsBuilder } from './bindings-wasi.js';
 
-import { default as wasmBase64 } from './autogen/wasm.js';
+import { default as wasmBase64, decompressedSize as wasmDecompressedSize } from './autogen/wasm.js';
 
 import { SmoldotWasmInstance } from './bindings.js';
 
@@ -51,11 +50,12 @@ export interface Config {
 }
 
 export async function startInstance(config: Config): Promise<SmoldotWasmInstance> {
-    // The actual Wasm bytecode is base64-decoded then gzip-decoded from a constant found in a
+    // The actual Wasm bytecode is base64-decoded then deflate-decoded from a constant found in a
     // different file.
     // This is suboptimal compared to using `instantiateStreaming`, but it is the most
     // cross-platform cross-bundler approach.
-    const wasmBytecode = pako.inflate(buffer.trustedBase64Decode(wasmBase64));
+    const wasmBytecode = new Uint8Array(wasmDecompressedSize)
+    tinyInflate(buffer.trustedBase64Decode(wasmBase64), wasmBytecode)
 
     let killAll: () => void;