--- a/bin/wasm-node/javascript/src/tiny-inflate/README.md
+++ b/bin/wasm-node/javascript/src/tiny-inflate/README.md
@@ -1,0 +1,3 @@
+This directory contains the `tiny-inflate` library, very slightly modified. Original found here: <https://github.com/foliojs/tiny-inflate>.
+
+Its source code is included within the source tree in order to avoid having to rely on the NPM repository.