cc https://github.com/paritytech/smoldot/issues/2427 https://github.com/paritytech/smoldot/issues/1802

Removes the `pako` library and uses `tiny-inflate` instead.

The objective of this PR is to remove a dependency, again with the idea of publishing to Deno at some point.
For this reason, the code of `tiny-inflate` (MIT) is directly embedded in the source tree. It is only 376 lines of code.

The `tiny-inflate` claims to be 30% to 50% slower than `pako`.
However, some quick benchmarking on my machine has shown the opposite.
With `pako`, I get 86ms to 88ms for base64 decoding + inflating.
With `tiny-inflate`, I get 78ms to 79ms instead.
The difference might be explained by the fact that I now use "raw" deflate, without any zlib header, as it's the only thing that `tiny-inflate` supports.